refactor: Change to Communication project net7.0

+++ build.ps1
@@ -1,6 +1,6 @@
 param(
     [string]$OutputPath = "",
-    [string]$Framework = "netcoreapp3.1",
+    [string]$Framework = "net7.0",
     [string]$KeyPath = "",
     [string]$LogPath = "",
     [switch]$Force
@@ -14,13 +14,20 @@ $propPaths = (
 )
 
 try {
-    $buildFile = "./build-temp.ps1"
     $solutionPath = Join-Path $PSScriptRoot $solutionPath -Resolve
     $propPaths = $propPaths | ForEach-Object { Join-Path $PSScriptRoot $_ -Resolve }
-    Invoke-WebRequest -Uri "https://raw.githubusercontent.com/s2quake/build/master/build.ps1" -OutFile $buildFile
+    if ($Env:COMMANDS_BUILD_PATH) {
+        $buildFile = $Env:COMMANDS_BUILD_PATH
+    }
+    else {
+        $buildFile = "./build-temp.ps1"
+        Invoke-WebRequest -Uri "https://raw.githubusercontent.com/s2quake/build/master/build.ps1" -OutFile $buildFile
+    }
     $buildFile = Resolve-Path $buildFile
     & $buildFile $solutionPath $propPaths -Publish -KeyPath $KeyPath -Sign -OutputPath $OutputPath -Framework $Framework -LogPath $LogPath -Force:$Force
 }
 finally {
-    Remove-Item $buildFile
+    if (-not $Env:COMMANDS_BUILD_PATH) {
+        Remove-Item $buildFile
+    }
 }
